make program its own page
